requirements.txt
streamlit
langchain
langchain-community
chromadb
unstructured
sentence-transformers
faiss-cpu
tiktoken
openai
pypdf
loguru
docx2txt
watchdog
python-dotenv
langchain_openai
pymilvus
weaviate-client
langchain-experimental